pyproject.toml
[project]
name = "restate-google-adk-example"
version = "0.1.0"
description = "A resilient agent using the Google ADK and Restate SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hypercorn>=0.17.3",
    "pydantic>=2.10.6",
    "restate-sdk[serde]>=0.12.0",
    "google-adk>=1.20.0",
    "litellm",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
restate-sdk = { path = "../sdk-python/dist/restate_sdk-0.12.0-cp313-cp313-linux_x86_64.whl" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
